requirements.txt
pandas
numpy
matplotlib
seaborn
ipython
python-dotenv
requests
scipy
optuna
scikit-learn
holidays
shap